requirements.txt
importlib-resources>=3.0.0
python-frontmatter>=1.0.0
pyyaml>=5.4.1
typing-extensions>=3.10.0.0
